Added Sample with tests
This WILL break the project, because the #include for the CUnit.h file uses a hardcoded filepath. To change this, unzip the CUnit-2.1-2.zip and manually put in the filepath for the CUnit.h file on YOUR machine. This will let it compile, and allow you to use CUnit's testing stuff.